SuperDepth3D 2.0.0 Update
SuperDepth3D 2.0.0
+Changed Texture Format to be kinder to GPUs
+Changed a Text Tool Tip

+++ Shaders/3DToElse.fx
@@ -101,8 +101,8 @@ sampler BackBuffer
 		Texture = BackBufferTex;
 	};
 	
-texture texCL  { Width = BUFFER_WIDTH; Height = BUFFER_HEIGHT; Format = RGBA32F;}; 
-texture texCR  { Width = BUFFER_WIDTH; Height = BUFFER_HEIGHT; Format = RGBA32F;}; 
+texture texCL  { Width = BUFFER_WIDTH; Height = BUFFER_HEIGHT; Format = RGBA8;}; 
+texture texCR  { Width = BUFFER_WIDTH; Height = BUFFER_HEIGHT; Format = RGBA8;}; 
 
 sampler SamplerCL
 	{
@@ -120,21 +120,21 @@ sampler SamplerCR
 		AddressW = BORDER;
 	};
 	
-texture texBB  { Width = BUFFER_WIDTH; Height = BUFFER_HEIGHT; Format = RGBA32F; MipLevels = 8;}; 
+texture texBB  { Width = BUFFER_WIDTH; Height = BUFFER_HEIGHT; Format = RGBA8; MipLevels = 8;}; 
 
 sampler SamplerBB
 	{
 		Texture = texBB;
 	};
   
- texture CurrentBackBuffer  { Width = BUFFER_WIDTH; Height = BUFFER_HEIGHT; Format = RGBA32F;}; 
+ texture CurrentBackBuffer  { Width = BUFFER_WIDTH; Height = BUFFER_HEIGHT; Format = RGBA8;}; 
 
 sampler CBackBuffer
 	{
 		Texture = CurrentBackBuffer;
 	};
 
-texture PastSingleBackBuffer  { Width = BUFFER_WIDTH; Height = BUFFER_HEIGHT; Format = RGBA32F;}; 
+texture PastSingleBackBuffer  { Width = BUFFER_WIDTH; Height = BUFFER_HEIGHT; Format = RGBA8;}; 
 
 sampler PSBackBuffer
 	{
